Core/Src: increase LoRaWAN FRMPayload size
The current implementation only support up to 16 bytes. The Ai and Si
block when doing FRMPayload encryption shouldn't be limited to one
block. This commit increase the LoRaWAN FRMPayload size that user
can submit to the encryption function (up to 255 bytes).

But the LoRa.c driver only support sending 255 bytes, which includes
header bytes (13) and FRMPayload. So the actual supported FRMPayload
size is 242 bytes.

Also stop using flags to indicate TIM4 completion status. Previously
the TIM4 period is used to signal 1 microseconds passed and raise a
flag. But that is not what we want, so remove it and let the TIM4
be a free-running counter with each peripheral clock tick equals 1
microseconds. So now we just need to check if the TIM4 counter
register matched the desired delay value.